Listen for dialog messages for Edge
The handler for Edge environments used to poll for localStorage, but
due to issues relating to Edge WebView2 rollout, localStorage is no
longer shared between the dialog prompt and add-in pane. This change
essentially reverts the original change from 2019 that updated the
handling to rely on localStorage.

The reason we are doing this is because:

1. The authentication flow is broken until MS resolves this bug:
https://github.com/OfficeDev/office-js/issues/1741

2. messageParent has always been the preferred communication medium
as noted in the DialogApi documentation:
https://docs.microsoft.com/en-us/office/dev/add-ins/develop/dialog-api-in-office-add-ins

3. There was a finding that the dialog needs to be opened with a URL
that matches the opener. With this change, messageParent now appears
to be working correctly.